First
Added files; main.c and makefile

Right now, this code is very early in development. Most of the
functions that will be used exist now, but are not yet written.

Important defines declared, functions thought about. Still need to do a
lot with interrupt service routines and writing the rest of the code.